Add TabTrayViewModel and move business logic there
